% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{calc.cdr}
\alias{calc.cdr}
\title{Calculate the cellular detection rate}
\usage{
calc.cdr(counts)
}
\arguments{
\item{counts}{A feature X cell table of raw counts.}
}
\value{
A vector of cellular detection rate per cell
}
\description{
Given a feature X cell table of raw counts, calculate the cellular detection
rate for each cell. The cellular detection rate of a cell is a number between
0 and 1 representing the fraction of the total number of features measured
that have a count > 0 in that cell. It is useful as a covariate in
differential expression analyses.
}
\examples{
cdr <- calc.cdr(seurat@assays$RNA@counts)
}
