% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{FindAllClusterDE}
\alias{FindAllClusterDE}
\title{Run differential expression analysis on every cluster}
\usage{
FindAllClusterDE(seurat, grouping.var, method = RunEdgeR)
}
\arguments{
\item{seurat}{A seurat object.}

\item{grouping.var}{The name of the metadata variable used to group the
samples into two groups to compare.}

\item{method}{A function that takes a seurat object, a cluster number, and a
grouping variable, and returns a \code{data.frame} with row names as
feature names, and the columns \code{p} and \code{fold.change}. See, for
example, \code{\link{RunEdgeR}}}
}
\value{
A \code{data.frame} containing one row for each cluster/feature
  combination, with columns \code{p}, \code{fold.change}, \code{cluster}, and
  \code{feature}.
}
\description{
Use a given function to perform differential expression analysis between two
conditions on every cluster in a Seurat object.
}
