% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helpers.R
\name{FindAllConservedMarkers}
\alias{FindAllConservedMarkers}
\title{Find conserved markers for each cluster}
\usage{
FindAllConservedMarkers(seurat, grouping.var, test.use = "wilcox")
}
\arguments{
\item{seurat}{A seurat object.}

\item{grouping.var}{The name of the metadata variable to use to separate the
cells into two conditions.}

\item{test.use}{String specifying the test to pass to \code{FindMarkers}.}
}
\value{
A \code{data.frame} of the outputs of all calls to
  \code{FindConservedMarkers}, row-bound together with the added columns
  \code{cluster} and \code{feature}.
}
\description{
Find conserved markers for each cluster. Basically, this does what
\link{\code{Seurat::FindAllMarkers}} does, except it uses
\link{\code{Seurat::FindConservedMarkers}} inside instead of plain old
\link{\code{Seurat::FindMarkers}}.
}
