% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{make.diffexp.heatmap}
\alias{make.diffexp.heatmap}
\title{Make a heatmap of the most differentially expressed genes in a cluster}
\usage{
make.diffexp.heatmap(
  seurat,
  diff.exp,
  cluster,
  grouping.var,
  min.fc = 1,
  max.p = 0.001,
  max.features = 30
)
}
\arguments{
\item{seurat}{A seurat object with defined clusters}

\item{diff.exp}{A table of differential expression p-values and fold changes
for all clusters, with columns \code{feature}, \code{cluster},
\code{fold.change}, and \code{p}.}

\item{cluster}{The number of the cluster to plot expression for}

\item{grouping.var}{The name of the metadata variable in \code{seurat} that
cells were grouped by in differential expression analysis}

\item{min.fc}{Minimum absolute value of fold change to include a feature in
heatmap}

\item{max.p}{Maximum differential expression p-value to include a feature in
heatmap}

\item{max.features}{Maximum number of features to include in the heatmap}
}
\value{
A ggplot object containing a heatmap with expression of at most
  \code{max.features} genes with the smallest differential expression
  p-values, ordered by fold change, for all cells in \code{cluster}, grouped
  by \code{grouping.var}.
}
\description{
Given a seurat object, a table of differential expression fold changes and
p-values, a cluster number, and a grouping variable name, make a heatmap
that shows per-cell expression of the genes with the smallest differential
expression p-values, ordered by fold change.
}
