% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{RunEdgeR}
\alias{RunEdgeR}
\title{Run edgeR to perform DE analysis}
\usage{
RunEdgeR(seurat, cluster, grouping.var)
}
\arguments{
\item{seurat}{A seurat object.}

\item{cluster}{The number of a cluster to perform DE analysis on}

\item{grouping.var}{The name of the metadata variable used to group the
samples into two groups to compare.}
}
\value{
A \code{data.frame} with row names as feature names, and the columns
  \code{p} and \code{fold.change}.
}
\description{
Run edgeR to assign a p-value to each feature for the null hypothesis that
this feature is not significantly differently expressed between the cells
of a given cluster between the two levels of a condition. This gives edgeR
raw counts, and uses cellular detection rate as a covariate. Based on
\link{https://github.com/csoneson/conquer_comparison/blob/master/scripts/apply_edgeRQLFDetRate.R}
}
\examples{
de.results <- RunEdgeR(seurat, 0, 'challenge_status')
}
