% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{RunWilcoxon}
\alias{RunWilcoxon}
\title{Run Wilcoxon tests to perform DE analysis.}
\usage{
RunWilcoxon(seurat, cluster, grouping.var)
}
\arguments{
\item{seurat}{A seurat object.}

\item{cluster}{The number of a cluster to perform DE analysis on}

\item{grouping.var}{The name of the metadata variable used to group the
samples into two groups to compare.}
}
\value{
A \code{data.frame} with row names as feature names, and the columns
  \code{p} and \code{fold.change}.
}
\description{
Use Wilcoxon tests to assign a p-value to each feature for the null
hypothesis that this feature is not significantly differently expressed in
the cells of a given cluster between two levels of a condition. This function
uses the CPM (counts per million) normalized expression levels as input to
the Wilcoxon test. Based on
\link{https://github.com/csoneson/conquer_comparison/blob/master/scripts/apply_Wilcoxon.R}
}
\examples{
de.results <- RunMAST(seurat, 0, 'challenge_status')
}
